requirements.txt
Flask
Werkzeug
torch==2.3.0 torchvision==0.18.0 torchaudio==2.3.0 --index-url https://download.pytorch.org/whl/cpu
opencv-python
numpy
scipy
pillow
pytesseract
ultralytics
vietocr
